Renamed erroneous fields

+++ src/Airslip.Common.Types/Transaction/TransactionTotalSummary.cs
@@ -2,9 +2,9 @@ namespace Airslip.Common.Types.Transaction;
 
 public class TransactionTotalSummary 
 {
-    public double? Total { get; set; }
-    public double? Subtotal { get; set; }
-    public double? Shipping { get; set; }
-    public double? Tax { get; set; }
-    public double? Discount { get; set; }
+    public long? Total { get; set; }
+    public long? Subtotal { get; set; }
+    public long? Shipping { get; set; }
+    public long? Tax { get; set; }
+    public long? Discount { get; set; }
 }
